refactor(shared): isSafeNodeId lives in shared so the renderer can use the SAME predicate
The renderer may not import src/core (measured: not one such import exists). It needs this exact
predicate to validate a project id before that id becomes a session-partition storage key. Moved,
not copied — core re-exports it, so every existing import path is untouched and a test pins that the
two names are one function object. The identity pin sits in core/remote-safety.test.ts because the
web typecheck program deliberately cannot see src/core — the very layering rule behind the move.

Mutations: copy-instead-of-re-export, drop the dot-segment refusal — 2/2 caught.
